+++ .github/CHANGELOG.md
@@ -150,6 +150,7 @@
 1. [A380X/FMS] Add STEP ALTs tab on VERT REV page & auto step climb functionality - @flogross89 (floridude)
 1. [A380X/FMS] Add selection of CLB/DES constraint if constraint type is unknown - @flogross89 (floridude)
 1. [FMS] Fix approach ident only showing on ND when approach is being flown - @BravoMike99 (bruno_pt99)
+1. [ND] Fix wind and speed display update rate for more frequent updates - @bulenteroglu (senolitam)
 
 ## 0.12.0